requirements.txt
absl-py==0.15.0
astunparse==1.6.3
blinker==1.8.2
cachetools==5.3.3
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.1.1
cycler==0.12.1
Flask==3.0.3
flatbuffers==1.12
fonttools==4.53.1
gast==0.3.3
google-auth==2.32.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==3.7
importlib_metadata==8.0.0
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
keras==2.15.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.7.5
numpy==1.23.4
oauthlib==3.2.2
opencv-python==4.10.0.84
opt-einsum==3.3.0
packaging==24.1
pandas==2.0.3
pillow==10.4.0
protobuf==3.20.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.15.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow @ file:///Users/charithesh/Downloads/plant%20seedlings%20classification/tensorflow-2.4.1-py3-none-any.whl
tensorflow-estimator==2.4.0
termcolor==1.1.0
typing-extensions==3.7.4.3
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.3
wrapt==1.12.1
zipp==3.19.2
